% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnout.R
\name{turnout_chart}
\alias{turnout_chart}
\title{A multi-line chart comparing multiple years of daily ballot returns}
\usage{
turnout_chart(year_type = "all", measure = "percent")
}
\arguments{
\item{measure}{The y axis measure to use, from \code{c("cumulative", "percent")}.}

\item{years}{Which collection of years to view, from \code{c("all", "local", "federal", "presidential", "midterm", "mayoral")}}
}
\description{
A multi-line chart comparing multiple years of daily ballot returns
}
